compareall.sh: Compare pairs against the unique
reference pdf created.

Besides, we don't need to do the comparisons twice now. This helps
us to reduce the consumed time by half.

Finally, incrase timeout to 240

This is the third change of a serie of changes to reduce the number
of time and memory required by the script when working with hundreds
of documents